Move put commands (p, P, gp, gP, ]p, [p) into their own file
